requirements.txt
asgiref==3.4.1
Django==3.2.8
django-braces==1.14.0
django-model-utils==4.2.0
django-star-ratings==0.9.1
django-taggit==1.5.1
gunicorn==20.1.0
Pillow==8.3.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
swapper==1.1.2.post1
whitenoise==5.3.0
psycopg2-binary==2.9.1